CMSIS-DSP : Testing framework improvements Added new tests for cfft. Improved error handling when not enough memory allocated in testmain.cpp Corrected arm_math.h (replaced inline by preprocessor definition). Added example debug script to read pattern and output of tests using scipy.

+++ CMSIS/DSP/Source/DistanceFunctions/arm_jensenshannon_distance_f32.c
@@ -30,7 +30,7 @@
 #include <math.h>
 
 
-static inline double rel_entr(double x, double y)
+static inline float32_t rel_entr(float32_t x, float32_t y)
 {
     return (x * log(x / y));
 }
